Echo handler accepting messages and replying
An approver was added, but things continued to not work. However,
after a make clean channels are now handled. The approver code
is commented out, and the handler still works.

The telepathy-bot.c file is now in a state where it's ready
to be refactored out into other files, and cleaned up for
actual development beyond a simple "get shit working" program