Feature parity with the proof of concept
The account and handler components now work the same
way as the proof of concept, but now with some gsignal
sauce which helps separate their concerns.

An local-xmpp account will be created if none exists
based on the BOTNAME env variable. The account is then
brought online and will start listening for text-channel
connections, responding with an uppercase echo of whatever
it receives.